requirements.txt
Flask
Flask-CORS
pymongo
python-dotenv
PyPDF2
requests
langchain-google-genai
gunicorn
